(refactor) move license file generation to library
